Update problem related permission
- user can create, edit problems and attachments in their course
- teacher and admin are not limited by above rule

+++ model/problem.py
@@ -61,7 +61,6 @@ def get_single_problem(user, problem):
 )
 @Request.doc('course', 'course', Course)
 @login_required
-@identity_verify(0, 1)
 def create_problem(
     user,
     title,
@@ -97,11 +96,13 @@ def create_problem(
 @problem_api.route('/<int:pid>', methods=['DELETE'])
 @Request.doc('pid', 'problem', Problem)
 @login_required
-@identity_verify(0, 1)
 def delete_problem(user, problem):
     '''
     delete a problem
     '''
+    # student can delete only self problem
+    if not problem.permission(user, {'w'}):
+        return HTTPError('Not enough permission', 403)
     problem.delete()
     return HTTPResponse(f'{problem} deleted.')
 
@@ -111,7 +112,6 @@ def delete_problem(user, problem):
 @Request.files('attachment')
 @Request.form('attachment_name')
 @login_required
-@identity_verify(0, 1)
 def patch_attachment(
     user,
     problem,
@@ -121,6 +121,8 @@ def patch_attachment(
     '''
     update the problem's attachment
     '''
+    if not problem.permission(user, {'w'}):
+        return HTTPError('Not enough permission', 403)
     if request.method == 'POST':
         try:
             problem.insert_attachment(
